+++ next.config.ts
@@ -16,6 +16,15 @@ const nextConfig: NextConfig = {
       ],
     }];
   },
+  async redirects() {
+    return [
+      {
+        source: "/find-booking",
+        destination: "/booking/find",
+        permanent: true,
+      },
+    ];
+  },
 };
 
 export default nextConfig;